LayoutTests/imported/w3c/web-platform-tests/html/rendering/replaced-elements/attributes-for-embedded-content-and-images/img-aspect-ratio-expected.txt

   

FAIL Image width and height attributes are used to infer aspect-ratio assert_approx_equals: expected 2.5 +/- 0.001 but got Infinity
